requirements.txt
aiogram>=3.0.0
tortoise-orm>=0.20.0
asyncpg>=0.27.0 